+++ packages/jest-worker/src/workers/ChildProcessWorker.ts
@@ -166,25 +166,15 @@ export default class ChildProcessWorker implements WorkerInterface {
     this._child.send(request);
   }
 
-  getWorkerId(): number {
+  getWorkerId() {
     return this._options.workerId;
   }
 
-  getStdout(): NodeJS.ReadableStream {
-    if (this._child.stdout === null) {
-      throw new Error(
-        'stdout is null - this should never happen. Please open up an issue at https://github.com/facebook/jest',
-      );
-    }
+  getStdout() {
     return this._child.stdout;
   }
 
-  getStderr(): NodeJS.ReadableStream {
-    if (this._child.stderr === null) {
-      throw new Error(
-        'stderr is null - this should never happen. Please open up an issue at https://github.com/facebook/jest',
-      );
-    }
+  getStderr() {
     return this._child.stderr;
   }
 }